Reorganized namespaces
